==> 76_Recommendation/01_Introduction to Recommendation Data.ipynb <==
import warnings
warnings.filterwarnings('ignore')

import numpy as np
import pandas as pd
import matplotlib.pyplot as plt

# Read in the datasets
movies = pd.read_csv('https://raw.githubusercontent.com/sidooms/MovieTweetings/master/latest/movies.dat', delimiter='::', header=None, names=['movie_id', 'movie', 'genre'], dtype={'movie_id': object}, engine='python', encoding='utf-8')
reviews = pd.read_csv('https://raw.githubusercontent.com/sidooms/MovieTweetings/master/latest/ratings.dat', delimiter='::', header=None, names=['user_id', 'movie_id', 'rating', 'timestamp'], dtype={'movie_id': object, 'user_id': object, 'timestamp': object}, engine='python', encoding='utf-8')

%matplotlib inline

movies.head()

reviews.head()

#### Explore Data

# Number of Movies
print(f"The number of movies: {movies.shape[0]}")

# Number of Ratings
print(f"The number of ratings: {reviews.shape[0]}")

reviews.isnull().sum()

# Number of Different genres
genres = []
for val in movies.genre:
    try:
        genres.extend(val.split('|'))
    except AttributeError:
        pass

genres = set(genres)
print(f"The number of genres is {len(genres)}")

reviews.describe()

#### Data Cleaning

# Pull Date if exists
create_date = lambda val: val[-5:-1] if val[-1] == ')' else np.nan

# Apply the Function to Pull the Date
movies['date'] = movies['movie'].apply(create_date)

movies.head()

# Return century of a movie as a dummy column
def add_movie_year(val):
    if val[:2] == yr:
        return 1
    else:
        return 0
    
# Apply Function
for yr in ['18', '19', '20']:
    movies[str(yr)+"00's"] = movies['date'].apply(add_movie_year)

movies.head()

# Function to Split and Return Values for Columns

def split_genres(val):
    try:
        if val.find(gene) > -1:
            return 1
        else:
            return 0
    except AttributeError:
        return 0
    
# Apply function to Each Genre
for gene in genres:
    movies[gene] = movies['genre'].apply(split_genres)

movies.head()

import datetime

change_timestamp = lambda val: datetime.datetime.fromtimestamp(int(val)).strftime('%Y-%m-%d %H:%M:%S')

reviews['date'] = reviews['timestamp'].apply(change_timestamp)

reviews.head()

# Now reviews and movies are the final dataframes with the necessary columns
reviews.to_csv('Data/reviews_clean.csv')
movies.to_csv('Data/movies_clean.csv')
