==> 76_Recommendation/02_Most Popular Recommendation.ipynb <==
import warnings
warnings.filterwarnings('ignore')

import numpy as np
import pandas as pd
import matplotlib.pyplot as plt

# Read in the Dataset
movies = pd.read_csv('Data/movies_clean.csv')
reviews = pd.read_csv('Data/reviews_clean.csv')

del movies['Unnamed: 0']
del reviews['Unnamed: 0']

movies.head()

reviews.head()

### How To Find The Most Popular Movies?
For this notebook, we have a single task. The task is that no matter the user, we need to provide a list of the recommendations based on simply the most popular items.

For this task, we will consider what is "most popular" based on the following criteria:

A movie with the highest average rating is considered best
With ties, movies that have more ratings are better
A movie must have a minimum of 5 ratings to be considered among the best movies
If movies are tied in their average rating and number of ratings, the ranking is determined by the movie that is the most recent rating
With these criteria, the goal for this notebook is to take a user_id and provide back the n_top recommendations.

def create_ranked_df(movies, reviews):
        '''
        INPUT
        movies - the movies dataframe
        reviews - the reviews dataframe
        
        OUTPUT
        ranked_movies - a dataframe with movies that are sorted by highest avg rating, more reviews, 
                        then time, and must have more than 4 ratings
        '''
        movie_ratings = reviews.groupby('movie_id')['rating']
        avg_ratings = movie_ratings.mean()
        num_ratings = movie_ratings.count()
        
        last_rating = pd.DataFrame(reviews.groupby('movie_id')['date'].max())
        last_rating.columns = ['last_rating']
        
        ratings_count_df = pd.DataFrame({'avg_rating': avg_ratings, 'num_rating': num_ratings})
        ratings_count_df = ratings_count_df.join(last_rating)
        
        movies_recs = movies.set_index('movie_id').join(ratings_count_df)
        
        ranked_movies = movies_recs.sort_values(['avg_rating', 'num_rating', 'last_rating'], ascending=False)
        
        ranked_movies = ranked_movies[ranked_movies['num_rating'] > 4]
        
        return ranked_movies

def popular_recommendations(user_id, n_top, ranked_movies):
    '''
    INPUT:
    user_id - the user_id (str) of the individual you are making recommendations for
    n_top - an integer of the number recommendations you want back
    ranked_movies - a pandas dataframe of the already ranked movies based on avg rating, count, and time

    OUTPUT:
    top_movies - a list of the n_top recommended movies by movie title in order best to worst'''
    
    top_movies = list(ranked_movies['movie'][:n_top])
    
    return top_movies

ranked_movies = create_ranked_df(movies, reviews)

ranked_movies

recs_20_for_1 = popular_recommendations('1', 20, ranked_movies)

recs_20_for_1

recs_20_for_1 = popular_recommendations('1', 20, ranked_movies)

recs_20_for_1

# Top 5 movies recommended for id 53968
recs_5_for_53968 = popular_recommendations('53968', 5, ranked_movies)
recs_5_for_53968

popular_recommendations('43', 35, ranked_movies)

def popular_recs_filtered(user_id, n_top, ranked_movies, years=None, genres=None):
    '''
    INPUT:
    user_id - the user_id (str) of the individual you are making recommendations for
    n_top - an integer of the number recommendations you want back
    ranked_movies - a pandas dataframe of the already ranked movies based on avg rating, count, and time
    years - a list of strings with years of movies
    genres - a list of strings with genres of movies
    
    OUTPUT:
    top_movies - a list of the n_top recommended movies by movie title in order best to worst
    '''
    # Filter movies based on year and genre
    if years is not None:
        ranked_movies = ranked_movies[ranked_movies['date'].isin(years)]

    if genres is not None:
        num_genre_match = ranked_movies[genres].sum(axis=1)
        ranked_movies = ranked_movies.loc[num_genre_match > 0, :]
            
    # create top movies list 
    top_movies = list(ranked_movies['movie'][:n_top])

    return top_movies

recs_20_for_1_filtered = popular_recs_filtered('1', 20, ranked_movies, years=['2015', '2016', '2017', '2018'], genres=['History'])

recs_20_for_1_filtered
